initial manually label videos

+++ server/ml-server/src/datalabeling/preparer.py
@@ -17,20 +17,25 @@ class Preparer():
         additional_body_datapoints = []
         # iterate over all files in folder
         for file in os.listdir(folder_path):
+            if file == "temp":
+                continue
             # grab video from temp folder and get information about it
             vid = util.grab_saved_video(os.path.join(folder_path, file))
-            # process video: downsample and get body datapoints
-            processed_video = self.vid_processor.process_video(vid, file, constants.FRAME_DIMENSION)
-            # delete temp video after processing
             if delete_original:
+                # process video: downsample and get body datapoints
+                processed_video = self.vid_processor.process_video_and_output_new_video(vid, file, constants.FRAME_DIMENSION)
+                # delete temp video after processing
                 os.remove(os.path.join(folder_path, file))
+            else:
+                processed_video = self.vid_processor.process_video(vid, file)
             # append to additional_body_datapoints
             additional_body_datapoints.append(processed_video.body_datapoints)
             # add to unlabeled videos 
             util.add_unlabeled_video(file, start_idx, unlabeled_dict)
             start_idx += 1
-        additional_body_datapoints_np_arr = np.array(additional_body_datapoints)
-        all_body_datapoints = np.concatenate((all_body_datapoints, additional_body_datapoints_np_arr), axis=0)
+        if len(additional_body_datapoints) > 0:
+            additional_body_datapoints_np_arr = np.array(additional_body_datapoints)
+            all_body_datapoints = np.concatenate((all_body_datapoints, additional_body_datapoints_np_arr), axis=0)
         return all_body_datapoints, unlabeled_dict
 
     def prepare_new_videos_for_labelling(self):
@@ -49,15 +54,17 @@ class Preparer():
 
     def prepare_all_videos_for_labelling(self):
         # create new empty data structures for preparing all videos
-        all_body_datapoints = util.create_new_all_body_datapoints()
         unlabeled_dict = util.create_new_unlabeled_dict()
+        all_body_datapoints = util.create_new_all_body_datapoints()
+        print(f"Before start preparer: len of unlabeled dict is {len(unlabeled_dict)}, len of all_body_datapoints is {len(all_body_datapoints)}, len of videos dir {len(os.listdir(constants.VIDEOS_DIR_PATH))}, len of temp dir is {len(os.listdir(os.path.join(constants.VIDEOS_DIR_PATH, "temp")))}")
         # reprocess all videos in videos folder
-        self.process_videos_in_folder(constants.VIDEOS_DIR_PATH, all_body_datapoints, unlabeled_dict, 0, False)
+        all_body_datapoints, unlabeled_dict = self.process_videos_in_folder(constants.VIDEOS_DIR_PATH, all_body_datapoints, unlabeled_dict, 0, False)
         # process all videos in temp folder
         start_idx = len(all_body_datapoints)
         temp_folder = os.path.join(constants.VIDEOS_DIR_PATH, "temp")
-        self.process_videos_in_folder(temp_folder, all_body_datapoints, unlabeled_dict, start_idx, True)
+        all_body_datapoints, unlabeled_dict = self.process_videos_in_folder(temp_folder, all_body_datapoints, unlabeled_dict, start_idx, True)
         # save body datapoints and unlabeled videos to file
+        print(f"After preparer: len of unlabeled dict is {len(unlabeled_dict)}, len of all_body_datapoints is {len(all_body_datapoints)}, len of videos dir {len(os.listdir(constants.VIDEOS_DIR_PATH))}, len of temp dir is {len(os.listdir(os.path.join(constants.VIDEOS_DIR_PATH, "temp")))}")
         util.save_all_body_datapoints(all_body_datapoints)
         util.save_unlabeled_dict(unlabeled_dict)
